[FIX] Specify size in kilobytes

+++ app/src/models/FilterValues.py
@@ -10,13 +10,13 @@ class FilterValues:
     """
 
     def __init__(self, name: str,
-                 size_from: Optional[int], size_to: Optional[int],
+                 size_from: Optional[float], size_to: Optional[float],
                  row_size_from: Optional[int], row_size_to: Optional[int],
                  column_size_from: Optional[int], column_size_to: Optional[int]):
         self.name: str = name.strip()
 
-        self.size_from: Optional[int] = size_from
-        self.size_to: Optional[int] = size_to
+        self.size_from: Optional[float] = size_from
+        self.size_to: Optional[float] = size_to
 
         self.row_size_from: Optional[int] = row_size_from
         self.row_size_to: Optional[int] = row_size_to
